Add files via upload
Jaro Winkler Distance indicates the similarity score between two Strings. The Jaro measure is the weighted sum of the percentage of matched characters from each file and transposed characters. Winkler increased this measure for matching initial characters.
This is a matrix containing the JW index of each commodity name against all others to ascertain which names can be grouped as a single commodity.